requirements.txt
streamlit
pandas
numpy
google-cloud-bigquery
db-dtypes
plotly
scipy
google-auth
google-auth-oauthlib
